# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OntoDB API",
    "description": "OntoDB — ontology-driven semantic multi-modal database. Supports SQL, SPARQL, vector search, and property graph queries.",
    "version": "0.1.0",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:7912",
      "description": "Local development server (default HTTP port)"
    }
  ],
  "paths": {
...
    "/api/vector/search": {
      "post": {
        "tags": ["Vector"],
        "summary": "Vector similarity search",
        "description": "Perform HNSW-based vector similarity search with optional SQL filter for hybrid queries.",
        "operationId": "vectorSearch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VectorSearchRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/QueryResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VectorSearchRequest": {
        "type": "object",
        "required": ["class", "column", "query_vector", "top_k"],
        "properties": {
          "class": { "type": "string", "description": "Target class (table) name", "example": "Product" },
          "column": { "type": "string", "description": "Vector column name", "example": "embedding" },
          "query_vector": { "type": "array", "items": { "type": "number", "format": "float" }, "description": "Query vector" },
          "top_k": { "type": "integer", "description": "Number of results to return", "example": 10 },
          "filter": { "type": "string", "description": "Optional SQL WHERE filter", "example": "price > 100" }
        }
      },
...
      "QueryResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "description": "Query result data" },
          "error": { "type": "string" },
          "elapsed_ms": { "type": "number", "format": "double" }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication. See config/api_keys.example.json for setup."
      }
    }
  },
  "security": [
    { "ApiKeyAuth": [] }
  ],
...
}